third_party/blink/web_tests/external/wpt/css/css-values/round-function-expected.txt

This is a testharness.js-based test.
Found 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] round(-1.5) should be used-value-equivalent to -1
  assert_equals: round(-1.5) and -1 serialize to the same thing in used values. expected "-1" but got "-2"
Harness: the test ran to completion.
